Reorganize and add more testcases
